Added delete method

+++ spring_rest_app/src/main/java/spring_rest/controller/MyRESTController.java
@@ -51,6 +51,11 @@ public class MyRESTController {
     // удаление работника
     @DeleteMapping("/employees/{id}")
     public String deleteEmployee(@PathVariable int id){
+        Employee employee = employeeService.getEmployee(id);
+
+        if (employee == null)
+            throw new NoSuchEmployeeException("There is no employee with id = " + id + " in database.");
+
         employeeService.deleteEmployee(id);
         return "Employee with id = " + id + " was deleted";
     }
